refactor: ContentFilterからStaticRejectClassifierを抽出
ContentFilterの静的検出ロジックをStaticRejectClassifierに分離し、
責務を明確にする。_is_fade_transitionの4つの独立条件ブロックも
個別メソッドに分割。
